Refactor ReviewsController and ReviewsService
- Update route for creating a review in ReviewsController to use "/create" endpoint instead of root endpoint.
- Add authorization check in ReviewsService to ensure only the creator of the review can delete it.

+++ server/src/services/ReviewsService.js
@@ -24,9 +24,9 @@ class ReviewsService {
   async deleteReview(reviewId, userInfo) {
     const reviewToBeDeleted = await dbContext.Review.findById(reviewId)
 
-    // if (userInfo.id != reviewToBeDeleted.creatorId) {
-    //   throw new Error('Unauthorized')
-    // }
+    if (userInfo.id != reviewToBeDeleted.creatorId) {
+      throw new Error('Unauthorized')
+    }
 
     await reviewToBeDeleted.deleteOne()
     return `Review titled: ${reviewToBeDeleted.title} has been deleted!`
